+++ components/Signin/useValidation.jsx
@@ -19,6 +19,8 @@ const schema = z.object({
     })
     .optional(),
   roleList: z.array(z.string()).min(1, '請選擇您的身份').optional(),
+  isSubscribeEmail: z.boolean().optional(),
+  interestList: z.array(z.string()).optional(),
 });
 
 const initialState = {
@@ -26,6 +28,7 @@ const initialState = {
   gender: '',
   roleList: [],
   isSubscribeEmail: true,
+  interestList: [],
 };
 
 const userReducer = (state, payload) => {
