+++ pages/signin/index.jsx
@@ -1,280 +1,84 @@
-import { useMemo, useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
-import { useSelector, useDispatch } from 'react-redux';
-import { createUser } from '@/redux/actions/user';
-import { GENDER, ROLE } from '@/constants/member';
-import dayjs from 'dayjs';
-import { Box, Typography, Button, Skeleton, TextField } from '@mui/material';
-import { LazyLoadImage } from 'react-lazy-load-image-component';
-import { MobileDatePicker } from '@mui/x-date-pickers/MobileDatePicker';
-import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
-import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
-import FormControlLabel from '@mui/material/FormControlLabel';
-import Checkbox from '@mui/material/Checkbox';
-import SEOConfig from '@/shared/components/SEO';
 import Navigation from '@/shared/components/Navigation_v2';
 import Footer from '@/shared/components/Footer_v2';
-import {
-  HomePageWrapper,
-  StyledContentWrapper,
-  StyledQuestionInput,
-} from '@/components/Signin/Signin.styled';
-import ErrorMessage from '@/components/Signin/ErrorMessage';
+import { HomePageWrapper } from '@/components/Signin/Signin.styled';
 import useProfileValidation from '@/components/Signin/useValidation';
-import { sendLoginConfirmation } from '@/utils/openLoginWindow';
+import { useAuth, useAuthDispatch } from '@/contexts/Auth';
+import Step1 from '@/components/Signin/Step1';
+import Step2 from '@/components/Signin/Step2';
+import TipModal from '@/components/Signin/Interest/TipModal';
 
 function SignInPage() {
   const router = useRouter();
-  const dispatch = useDispatch();
-
-  const { id, token } = router.query;
+  const [step, setStep] = useState(1);
+  const [open, setOpen] = useState(false);
   const { errors, onChangeHandler, userState, validateFields } =
     useProfileValidation();
 
-  const { createdDate, updatedDate } = useSelector((state) => state?.user);
+  const { isLoggedIn, token } = useAuth();
+  const { updateUser } = useAuthDispatch();
 
-  // Oath login
-  useEffect(() => {
-    sendLoginConfirmation(id, token, `/signin?id=${id}`);
-  }, [id, token]);
-
-  useEffect(() => {
-    if (createdDate !== updatedDate) {
-      router.push('/profile');
-    }
-  }, [createdDate, updatedDate]);
-
-  const handleRoleListChange = (value) => {
-    const { roleList } = userState;
-    const updatedRoleList = roleList.includes(value)
-      ? roleList.filter((role) => role !== value)
-      : [...roleList, value];
-    onChangeHandler({ key: 'roleList', value: updatedRoleList });
-  };
-
-  const onCreateUser = () => {
-    const { birthDay, gender, roleList, isSubscribeEmail } = userState;
-    if (validateFields({ birthDay, gender, roleList }, true)) {
+  const handleSubmit = async () => {
+    if (validateFields(userState, true)) {
       const payload = {
-        id,
-        birthDay: birthDay.toISOString(),
-        gender,
-        roleList,
-        isSubscribeEmail,
+        ...userState,
+        isSendEmail: true,
+        birthDay: userState.birthDay.toISOString(),
       };
-      dispatch(createUser(payload));
-      router.push(`/signin/interest?id=${id}`);
+      try {
+        await updateUser(payload);
+      } catch (error) {
+        console.error(error);
+      }
     }
   };
 
-  const SEOData = useMemo(
-    () => ({
-      title: '編輯我的島島資料｜島島阿學',
-      description:
-        '「島島阿學」盼能透過建立多元的學習資源網絡，讓自主學習者能找到合適的成長方法，進一步成為自己想成為的人，從中培養共好精神。目前正積極打造「可共編的學習資源平台」。',
-      keywords: '島島阿學',
-      author: '島島阿學',
-      copyright: '島島阿學',
-      imgLink: 'https://www.daoedu.tw/preview.webp',
-      link: `${process.env.HOSTNAME}${router?.asPath}`,
-    }),
-    [router?.asPath],
-  );
+  useEffect(() => {
+    if (isLoggedIn) {
+      setOpen(true);
+    } else if (!token) {
+      router.push('/');
+    }
+  }, [isLoggedIn, token]);
 
   return (
     <>
-      <SEOConfig data={SEOData} />
-      <Box>
-        <LocalizationProvider dateAdapter={AdapterDayjs}>
-          <Box sx={{ minHeight: '100vh' }}>
-            <StyledContentWrapper>
-              <h2>基本資料</h2>
-              <Box sx={{ marginTop: '24px', width: '100%', padding: '0 5%' }}>
-                <StyledQuestionInput>
-                  <Typography>生日 *</Typography>
-                  <MobileDatePicker
-                    maxDate={dayjs().subtract(16, 'year')}
-                    defaultCalendarMonth={dayjs().subtract(18, 'year')}
-                    label="birthDay"
-                    inputFormat="YYYY/MM/DD"
-                    value={userState.birthDay}
-                    onChange={(date) =>
-                      onChangeHandler({ key: 'birthDay', value: date })
-                    }
-                    renderInput={(params) => (
-                      <TextField {...params} sx={{ width: '100%' }} label="" />
-                    )}
-                  />
-                  <ErrorMessage errText={errors.birthDay} />
-                </StyledQuestionInput>
-                <StyledQuestionInput>
-                  <Typography>性別 *</Typography>
-                  <Box
-                    sx={{
-                      display: 'flex',
-                      justifyContent: 'space-between',
-                      alignItems: 'center',
-                      width: '100%',
-                    }}
-                  >
-                    {GENDER.map(({ label, value }) => (
-                      <Box
-                        key={label}
-                        onClick={() =>
-                          onChangeHandler({ key: 'gender', value })
-                        }
-                        sx={{
-                          border: '1px solid #DBDBDB',
-                          borderRadius: '8px',
-                          padding: '10px',
-                          width: 'calc(calc(100% - 16px) / 3)',
-                          display: 'flex',
-                          justifyItems: 'center',
-                          alignItems: 'center',
-                          cursor: 'pointer',
-                          ...(userState.gender === value
-                            ? {
-                                backgroundColor: '#DEF5F5',
-                                border: '1px solid #16B9B3',
-                              }
-                            : {}),
-                        }}
-                      >
-                        <Typography sx={{ margin: 'auto' }}>{label}</Typography>
-                      </Box>
-                    ))}
-                  </Box>
-                  <ErrorMessage errText={errors.gender} />
-                </StyledQuestionInput>
-                <StyledQuestionInput>
-                  <Typography>身份 *</Typography>
-                  <Box
-                    sx={{
-                      display: 'flex',
-                      flexWrap: 'wrap',
-                      justifyContent: 'space-between',
-                      alignItems: 'center',
-                      width: '100%',
-                      marginTop: '10px',
-                    }}
-                  >
-                    {ROLE.map(({ label, value, image }) => (
-                      <Box
-                        key={label}
-                        onClick={() => handleRoleListChange(value)}
-                        sx={{
-                          border: '1px solid #DBDBDB',
-                          borderRadius: '8px',
-                          padding: '10px',
-                          margin: '4px',
-                          width: 'calc(calc(100% - 24px) / 3)',
-                          flexBasis: 'calc(calc(100% - 24px) / 3)',
-                          display: 'flex',
-                          flexDirection: 'column',
-                          justifyItems: 'center',
-                          alignItems: 'center',
-                          cursor: 'pointer',
-                          ...(userState.roleList.includes(value)
-                            ? {
-                                backgroundColor: '#DEF5F5',
-                                border: '1px solid #16B9B3',
-                              }
-                            : {}),
-                          '@media (max-width: 767px)': {
-                            height: '100% auto',
-                            width: 'calc(calc(100% - 24px) / 2)',
-                            flexBasis: 'calc(calc(100% - 24px) / 2)',
-                          },
-                        }}
-                      >
-                        <LazyLoadImage
-                          alt={label}
-                          src={image}
-                          effect="opacity"
-                          style={{
-                            height: '100px',
-                            width: '100%',
-                            borderRadius: '6px',
-                            // background: 'rgba(240, 240, 240, .8)',
-                            objectFit: 'cover',
-                            objectPosition: 'center',
-                            '@media (maxWidth: 767px)': {
-                              width: '100%',
-                            },
-                          }}
-                          placeholder={
-                            // eslint-disable-next-line react/jsx-wrap-multilines
-                            <Skeleton
-                              sx={{
-                                height: '100px',
-                                width: '100%',
-                                borderRadius: '6px',
-                                background: 'rgba(240, 240, 240, .8)',
-                                marginTop: '4px',
-                              }}
-                              variant="rectangular"
-                              animation="wave"
-                            />
-                          }
-                        />
-                        <Typography
-                          sx={{
-                            margin: 'auto',
-                            marginTop: '10px',
-                            ...(userState.roleList.includes(value)
-                              ? {
-                                  fontWeight: 700,
-                                }
-                              : {}),
-                          }}
-                        >
-                          {label}
-                        </Typography>
-                      </Box>
-                    ))}
-                  </Box>
-                  <ErrorMessage errText={errors.roleList} />
-                </StyledQuestionInput>
-                <FormControlLabel
-                  sx={{
-                    marginTop: '20px',
-                  }}
-                  control={
-                    // eslint-disable-next-line react/jsx-wrap-multilines
-                    <Checkbox
-                      checked={userState.isSubscribeEmail}
-                      onChange={(event) =>
-                        onChangeHandler({
-                          key: 'isSubscribeEmail',
-                          value: event.target.checked,
-                        })
-                      }
-                    />
-                  }
-                  label="訂閱電子報與島島阿學的新資訊"
-                />
-                <Button
-                  sx={{
-                    width: '100%',
-                    height: '40px',
-                    borderRadius: '20px',
-                    margin: '24px 0px 45px 0px',
-                    color: '#ffff',
-                    bgcolor: '#16B9B3',
-                  }}
-                  variant="contained"
-                  onClick={onCreateUser}
-                >
-                  下一步
-                </Button>
-                {Object.values(errors).join('') && (
-                  <ErrorMessage errText="請將資訊填寫完整" />
-                )}
-              </Box>
-            </StyledContentWrapper>
-          </Box>
-        </LocalizationProvider>
-      </Box>
+      {step === 1 && (
+        <Step1
+          errors={errors}
+          onChangeHandler={onChangeHandler}
+          userState={userState}
+          onNext={() => {
+            if (validateFields(userState, true)) {
+              setStep(2);
+              window.scrollTo({ top: 0, behavior: 'smooth' });
+            }
+          }}
+        />
+      )}
+      {step === 2 && (
+        <Step2
+          onChangeHandler={onChangeHandler}
+          userState={userState}
+          onBack={() => {
+            setStep(1);
+            window.scrollTo({ top: 0, behavior: 'smooth' });
+          }}
+          onNext={handleSubmit}
+        />
+      )}
+      <TipModal
+        open={open}
+        onClose={() => {
+          setOpen(false);
+          router.replace('/');
+        }}
+        onOk={() => {
+          setOpen(false);
+          router.replace('/profile');
+        }}
+      />
     </>
   );
 }
